feat: Use annotation instead of abstract class

+++ controller/src/main/java/com/example/demo/controller/ShoeController.java
@@ -3,7 +3,6 @@ package com.example.demo.controller;
 import com.example.demo.dto.in.ShoeFilter;
 import com.example.demo.dto.out.Shoes;
 import com.example.demo.facade.ShoeFacade;
-import java.math.BigInteger;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -19,7 +18,7 @@ public class ShoeController {
   private final ShoeFacade shoeFacade;
 
   @GetMapping(path = "/search")
-  public ResponseEntity<Shoes> all(ShoeFilter filter, @RequestHeader BigInteger version){
+  public ResponseEntity<Shoes> all(ShoeFilter filter, @RequestHeader Integer version){
 
     return ResponseEntity.ok(shoeFacade.get(version).search(filter));
